+++ packages/notification-center/src/components/notification-center/NotificationCenter.tsx
@@ -3,8 +3,10 @@ import { IMessage, IMessageAction, ButtonTypeEnum } from '@novu/shared';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { AppContent } from './components';
 import { useNovuContext } from '../../hooks';
-import { INovuThemeProvider, NovuThemeProvider, NotificationCenterContext } from '../../store';
-import { ColorScheme, ITab, ListItem } from '../../index';
+import { NotificationCenterContext } from '../../store/notification-center.context';
+import { ITab, ListItem } from '../../shared/interfaces';
+import { ColorScheme } from '../../shared/config/colors';
+import { INovuThemeProvider, NovuThemeProvider } from '../../store/novu-theme-provider.context';
 
 export interface INotificationCenterProps {
   onUrlChange?: (url: string) => void;
