+++ packages/notification-center/src/components/notification-center/components/Main.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import { FeedsTabs } from './FeedsTabs';
-import { ITab } from '../../../index';
 
 export function Main() {
   return (
