+++ packages/notification-center/src/components/notification-center/components/layout/tabs/Tabs.styles.ts
@@ -1,4 +1,4 @@
-import { Badge, createStyles, MantineTheme } from '@mantine/core';
+import { createStyles, MantineTheme } from '@mantine/core';
 import { INovuTheme } from '../../../../../store/novu-theme.context';
 import { ICommonTheme } from '../../../../../store/novu-theme-provider.context';
 import { colors } from '../../../../../shared/config/colors';
