+++ packages/notification-center/src/hooks/use-auth.hook.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { AuthContext } from '../store/auth.context';
-import { IAuthContext } from '../index';
+import { IAuthContext } from '../shared/interfaces';
 
 export function useAuth() {
   const { token, applyToken, user, setUser, isLoggedIn } = useContext<IAuthContext>(AuthContext);
