+++ packages/notification-center/src/hooks/use-socket.hook.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { SocketContext } from '../store/socket/socket.store';
-import { ISocketContext } from '../index';
+import { ISocketContext } from '../shared/interfaces';
 
 export function useSocket() {
   const { socket } = useContext<ISocketContext>(SocketContext);
