+++ packages/notification-center/src/hooks/use-unseen-count.hook.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { UnseenCountContext } from '../store';
+import { UnseenCountContext } from '../store/unseen-count.context';
 
 export function useUnseenCount() {
   const { unseenCount, setUnseenCount } = useContext(UnseenCountContext);
