+++ packages/notification-center/src/shared/config/colors.ts
@@ -17,3 +17,5 @@ export const colors = {
   vertical: 'linear-gradient(0deg,#FF512F 0%,#DD2476 100%)',
   horizontal: 'linear-gradient(99deg,#DD2476 0% 0%,#FF512F 100% 100%)',
 };
+
+export declare type ColorScheme = 'light' | 'dark';
