+++ packages/notification-center/src/shared/enums/screens.enum.ts
@@ -1,0 +1,4 @@
+export enum ScreensEnum {
+  NOTIFICATIONS = 'notifications',
+  SETTINGS = 'settings',
+}
