--- packages/notification-center/src/store/index.ts
@@ -1,16 +1,0 @@
-export * from './notifications-provider.context';
-export * from './auth-provider.context';
-export * from './novu-provider.context';
-export * from './screens-provider.context';
-export * from './novu-theme-provider.context';
-export * from './unseen-provider.context';
-export * from './socket-initialization-provider.context';
-
-export * from './unseen-count.context';
-export * from './screens.context';
-export * from './novu-theme.context';
-export * from './i18n.context';
-export * from './api.context';
-export * from './auth.context';
-export * from './notification-center.context';
-export * from './notifications.context';
