+++ packages/notification-center/src/store/notification-center.context.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { IMessage, ButtonTypeEnum } from '@novu/shared';
-import { INotificationCenterContext, ITab } from '../index';
+import { INotificationCenterContext, ITab } from '../shared/interfaces';
 
 export const NotificationCenterContext = React.createContext<INotificationCenterContext>({
   onUrlChange: (url: string) => {},
