+++ packages/notification-center/src/store/novu-provider.context.ts
@@ -1,4 +1,4 @@
 import React from 'react';
-import { INovuProviderContext } from '../index';
+import { INovuProviderContext } from '../shared/interfaces';
 
 export const NovuContext = React.createContext<INovuProviderContext | null>(null);
